+++ pgdrive/manager/agent_manager.py
@@ -7,7 +7,7 @@ from pgdrive.policy.AI_protect_policy import AIProtectPolicy
 import logging
 from typing import Dict
 
-from gym.spaces import Box, Dict
+from gym.spaces import Box, Dict, MultiDiscrete
 
 from pgdrive.manager.base_manager import BaseManager
 
@@ -128,7 +128,7 @@ class AgentManager(BaseManager):
                 assert isinstance(obs_space, Dict), "Multi-agent observation should be gym.Dict"
             action_space = self._init_action_spaces[agent_id]
             self.action_spaces[vehicle.name] = action_space
-            assert isinstance(action_space, Box)
+            assert isinstance(action_space, Box) or isinstance(action_space, MultiDiscrete)
         self.next_agent_count = len(init_vehicles)
 
     def finish(self, agent_name, ignore_delay_done=False):
